+++ R/MONTHLY_LEVELS.R
@@ -12,8 +12,8 @@
 
 #' Extract monthly levels information from the HYDAT database
 #'
-#' Provides wrapper to turn the hy_monthly_levels table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
-#'   \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
+#' Provides wrapper to turn the hy_monthly_levels table in HYDAT into a tidy data frame.  \code{station_number} and
+#'   \code{prov_terr_state_loc} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
 #'   That is a large vector for \code{hy_monthly_levels}.
 #'
 #' @inheritParams hy_stations
@@ -25,10 +25,10 @@
 #'
 #' @examples
 #' \donttest{
-#' hy_monthly_levels(STATION_NUMBER = c("02JE013","08MF005"), 
+#' hy_monthly_levels(station_number = c("02JE013","08MF005"), 
 #'   start_date = "1996-01-01", end_date = "2000-01-01")
 #'
-#' hy_monthly_levels(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
+#' hy_monthly_levels(prov_terr_state_loc = "PE", hydat_path = "H:/Hydat.sqlite3")
 #'           }
 #' @family HYDAT functions
 #' @source HYDAT
@@ -36,11 +36,11 @@
 
 
 
-hy_monthly_levels <- function(STATION_NUMBER = NULL,
+hy_monthly_levels <- function(station_number = NULL,
                            hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"),
-                           PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
-  if (!is.null(STATION_NUMBER) && STATION_NUMBER == "ALL") {
-    stop("Deprecated behaviour.Omit the STATION_NUMBER = \"ALL\" argument. See ?hy_monthly_levels for examples.")
+                           prov_terr_state_loc = NULL, start_date ="ALL", end_date = "ALL") {
+  if (!is.null(station_number) && station_number == "ALL") {
+    stop("Deprecated behaviour.Omit the station_number = \"ALL\" argument. See ?hy_monthly_levels for examples.")
   }
   
   ## Check if hydat is present
@@ -81,7 +81,7 @@ hy_monthly_levels <- function(STATION_NUMBER = NULL,
   on.exit(DBI::dbDisconnect(hydat_con))
 
   ## Determine which stations we are querying
-  stns <- station_choice(hydat_con, STATION_NUMBER, PROV_TERR_STATE_LOC)
+  stns <- station_choice(hydat_con, station_number, prov_terr_state_loc)
 
 
   ## Data manipulations to make it "tidy"