+++ R/SED_DLY_LOADS.R
@@ -12,8 +12,8 @@
 
 #' Extract daily sediment load information from the HYDAT database
 #'
-#' Provides wrapper to turn the hy_sed_daily_loads table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
-#' \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
+#' Provides wrapper to turn the hy_sed_daily_loads table in HYDAT into a tidy data frame.  \code{station_number} and
+#' \code{prov_terr_state_loc} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
 #' That is a large vector for \code{hy_sed_daily_loads}.
 #'
 #' @inheritParams hy_stations
@@ -24,7 +24,7 @@
 #'
 #' @examples
 #' \donttest{
-#' hy_sed_daily_loads(PROV_TERR_STATE_LOC = "PE", hydat_path = "H:/Hydat.sqlite3")
+#' hy_sed_daily_loads(prov_terr_state_loc = "PE", hydat_path = "H:/Hydat.sqlite3")
 #'           }
 #'
 #' @family HYDAT functions
@@ -33,9 +33,9 @@
 
 
 
-hy_sed_daily_loads <- function(STATION_NUMBER = NULL, 
+hy_sed_daily_loads <- function(station_number = NULL, 
                           hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
-                          PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
+                          prov_terr_state_loc = NULL, start_date ="ALL", end_date = "ALL") {
   
   ## Check if hydat is present
   if (!file.exists(hydat_path)){
@@ -76,7 +76,7 @@ hy_sed_daily_loads <- function(STATION_NUMBER = NULL,
   on.exit(DBI::dbDisconnect(hydat_con))
 
   ## Determine which stations we are querying
-  stns <- station_choice(hydat_con, STATION_NUMBER, PROV_TERR_STATE_LOC)
+  stns <- station_choice(hydat_con, station_number, prov_terr_state_loc)
 
   ## Data manipulations
   sed_dly_loads <- dplyr::tbl(hydat_con, "SED_DLY_LOADS")