+++ R/SED_MONTHLY_LOADS.R
@@ -12,8 +12,8 @@
 
 #' Extract monthly flows information from the HYDAT database
 #'
-#' Provides wrapper to turn the hy_sed_monthly_loads table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
-#'   \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
+#' Provides wrapper to turn the hy_sed_monthly_loads table in HYDAT into a tidy data frame.  \code{station_number} and
+#'   \code{prov_terr_state_loc} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
 #'   That is a large vector for \code{hy_sed_monthly_loads}.
 #'
 #' @inheritParams hy_stations
@@ -26,7 +26,7 @@
 #' @examples
 #' \donttest{
 #'
-#' hy_sed_daily_suscon(STATION_NUMBER = "01CE003")
+#' hy_sed_daily_suscon(station_number = "01CE003")
 #'           }
 #' @family HYDAT functions
 #' @source HYDAT
@@ -34,11 +34,11 @@
 
 
 
-hy_sed_monthly_loads <- function(STATION_NUMBER = NULL,
+hy_sed_monthly_loads <- function(station_number = NULL,
                               hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
-                              PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
-  if (!is.null(STATION_NUMBER) && STATION_NUMBER == "ALL") {
-    stop("Deprecated behaviour.Omit the STATION_NUMBER = \"ALL\" argument. See ?hy_sed_monthly_loads for examples.")
+                              prov_terr_state_loc = NULL, start_date ="ALL", end_date = "ALL") {
+  if (!is.null(station_number) && station_number == "ALL") {
+    stop("Deprecated behaviour.Omit the station_number = \"ALL\" argument. See ?hy_sed_monthly_loads for examples.")
   }
   
   if (start_date == "ALL" & end_date == "ALL") {
@@ -73,7 +73,7 @@ hy_sed_monthly_loads <- function(STATION_NUMBER = NULL,
   on.exit(DBI::dbDisconnect(hydat_con))
   
   ## Determine which stations we are querying
-  stns <- station_choice(hydat_con, STATION_NUMBER, PROV_TERR_STATE_LOC)
+  stns <- station_choice(hydat_con, station_number, prov_terr_state_loc)
   
   
   ## Data manipulations to make it "tidy"