+++ R/SED_MONTHLY_SUSCON.R
@@ -12,8 +12,8 @@
 
 #' Extract monthly flows information from the HYDAT database
 #'
-#' Provides wrapper to turn the hy_sed_monthly_suscon table in HYDAT into a tidy data frame.  \code{STATION_NUMBER} and
-#'   \code{PROV_TERR_STATE_LOC} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
+#' Provides wrapper to turn the hy_sed_monthly_suscon table in HYDAT into a tidy data frame.  \code{station_number} and
+#'   \code{prov_terr_state_loc} can both be supplied. If both are omitted all values from the \code{hy_stations} table are returned.
 #'   That is a large vector for \code{hy_sed_monthly_suscon}.
 #'
 #' @inheritParams hy_stations
@@ -25,7 +25,7 @@
 #'
 #' @examples
 #' \donttest{
-#' hy_sed_monthly_suscon(STATION_NUMBER = "08MF005")
+#' hy_sed_monthly_suscon(station_number = "08MF005")
 #'           }
 #' @family HYDAT functions
 #' @source HYDAT
@@ -33,11 +33,11 @@
 
 
 
-hy_sed_monthly_suscon <- function(STATION_NUMBER = NULL, 
+hy_sed_monthly_suscon <- function(station_number = NULL, 
                                hydat_path = paste0(rappdirs::user_data_dir(),"\\Hydat.sqlite3"), 
-                               PROV_TERR_STATE_LOC = NULL, start_date ="ALL", end_date = "ALL") {
-  if (!is.null(STATION_NUMBER) && STATION_NUMBER == "ALL") {
-    stop("Deprecated behaviour.Omit the STATION_NUMBER = \"ALL\" argument. See ?hy_sed_monthly_suscon for examples.")
+                               prov_terr_state_loc = NULL, start_date ="ALL", end_date = "ALL") {
+  if (!is.null(station_number) && station_number == "ALL") {
+    stop("Deprecated behaviour.Omit the station_number = \"ALL\" argument. See ?hy_sed_monthly_suscon for examples.")
   }
   
   ## Check if hydat is present
@@ -77,7 +77,7 @@ hy_sed_monthly_suscon <- function(STATION_NUMBER = NULL,
   on.exit(DBI::dbDisconnect(hydat_con))
 
   ## Determine which stations we are querying
-  stns <- station_choice(hydat_con, STATION_NUMBER, PROV_TERR_STATE_LOC)
+  stns <- station_choice(hydat_con, station_number, prov_terr_state_loc)
 
 
   ## Data manipulations to make it "tidy"