+++ R/internal.R
@@ -8,35 +8,35 @@
 #' @keywords internal
 #'
 #'
-station_choice <- function(hydat_con, STATION_NUMBER, PROV_TERR_STATE_LOC) {
+station_choice <- function(hydat_con, station_number, prov_terr_state_loc) {
 
 
-  ## Only possible values for PROV_TERR_STATE_LOC
+  ## Only possible values for prov_terr_state_loc
   stn_option <- dplyr::tbl(hydat_con, "STATIONS") %>%
     dplyr::distinct(PROV_TERR_STATE_LOC) %>%
     dplyr::pull(PROV_TERR_STATE_LOC)
 
-  ## If not STATION_NUMBER arg is supplied then this controls how to handle the PROV arg
-  if ((is.null(STATION_NUMBER) & !is.null(PROV_TERR_STATE_LOC))) {
-    STATION_NUMBER <- "ALL" ## All stations
-    prov <- PROV_TERR_STATE_LOC ## Prov info
+  ## If not station_number arg is supplied then this controls how to handle the PROV arg
+  if ((is.null(station_number) & !is.null(prov_terr_state_loc))) {
+    station_number <- "ALL" ## All stations
+    prov <- prov_terr_state_loc ## Prov info
 
     if (any(!prov %in% stn_option) == TRUE) {
-      stop("Invalid PROV_TERR_STATE_LOC value")
+      stop("Invalid prov_terr_state_loc value")
     }
   }
 
-  ## If PROV arg is supplied then simply use the STATION_NUMBER independent of PROV
-  if (is.null(PROV_TERR_STATE_LOC)) {
-    STATION_NUMBER <- STATION_NUMBER
+  ## If PROV arg is supplied then simply use the station_number independent of PROV
+  if (is.null(prov_terr_state_loc)) {
+    station_number <- station_number
   }
 
 
   ## Steps to create the station vector
-  stns <- STATION_NUMBER
+  stns <- station_number
 
   ## Get all stations
-  if (is.null(stns) == TRUE && is.null(PROV_TERR_STATE_LOC) == TRUE) {
+  if (is.null(stns) == TRUE && is.null(prov_terr_state_loc) == TRUE) {
     stns <- dplyr::tbl(hydat_con, "STATIONS") %>%
       dplyr::collect() %>%
       dplyr::pull(STATION_NUMBER)